Move group() to reaction.py
